Update tutorial-managed-virtual-network-sql-managed-instance.md
Recently, while working with a customer, I discovered that IP forwarding was set to temporary. After a reboot, the connection was not working. To resolve this issue, I added the code from the ip_fwd.sh script and also instructed the customer to manually add this to the config file.